backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.5
opensearch-py==2.8.0
docling>=2.0.0
easyocr>=1.7.0
pydantic-settings==2.10.1
python-dotenv==1.1.1
